// src/cars/interface/cars.interface.ts
export interface ICars {
    id: string;
    name?: string;
    marka?: string,
    tanirovka?: boolean;
    motor?: string;
    year?: string;
    color?: string;
    distance?: string;
    gearbook?: string;
    description?: string;
    narxi?: string;
    main_img?: string;
    tashqi_img?: string;
    ichki_img?: string;
}